+++ Buffer.cpp
@@ -1,21 +1,21 @@
 #include "Buffer.h"
 
-Buffer::Buffer(size_t width, size_t height) {
+Buffer::Buffer(int width, int height) {
 	this->width = width;
 	this->height = height;
 	this->data = new uint32_t[width * height];
 }
 
 void Buffer::colorClear(uint32_t color) {
-	for (size_t i = 0; i < this->width * this->height; ++i)
+	for (int i = 0; i < this->width * this->height; ++i)
 	{
 		this->data[i] = color;
 	}
 }
 
-void Buffer::drawText(const Sprite& textSpritesheet, const char* text, size_t x, size_t y, uint32_t color) {
-	size_t xp = x;
-	size_t stride = textSpritesheet.width * textSpritesheet.height;
+void Buffer::drawText(const Sprite& textSpritesheet, const char* text, int x, int y, uint32_t color) {
+	int xp = x;
+	int stride = textSpritesheet.width * textSpritesheet.height;
 	Sprite sprite = textSpritesheet;
 	for (const char* charp = text; *charp != '\0'; ++charp)
 	{
@@ -27,7 +27,7 @@ void Buffer::drawText(const Sprite& textSpritesheet, const char* text, size_t x,
 	}
 }
 
-void Buffer::drawSprite(const Sprite* sprite, size_t startingX, size_t startingY, uint32_t color) {
+void Buffer::drawSprite(const Sprite* sprite, int startingX, int startingY, uint32_t color) {
 	for (int x = 0; x < sprite->width; ++x) {
 		for (int y = 0; y < sprite->height; ++y) {
 			if (sprite->data[y * sprite->width + x] &&
@@ -39,18 +39,18 @@ void Buffer::drawSprite(const Sprite* sprite, size_t startingX, size_t startingY
 	}
 }
 
-void Buffer::drawNumber(const Sprite& numberSpritesheet, size_t number, size_t x, size_t y, uint32_t color) {
+void Buffer::drawNumber(const Sprite& numberSpritesheet, int number, int x, int y, uint32_t color) {
 	uint8_t digits[64];
-	size_t numDigits = 0;
-	size_t currentNumber = number;
+	int numDigits = 0;
+	int currentNumber = number;
 	do {
 		digits[numDigits++] = currentNumber % 10;
 		currentNumber = currentNumber / 10;
 	} while (currentNumber > 0);
-	size_t xp = x;
-	size_t stride = numberSpritesheet.width * numberSpritesheet.height;
+	int xp = x;
+	int stride = numberSpritesheet.width * numberSpritesheet.height;
 	Sprite sprite = numberSpritesheet;
-	for (size_t i = 0; i < numDigits; ++i) {
+	for (int i = 0; i < numDigits; ++i) {
 		uint8_t digit = digits[numDigits - i - 1];
 		sprite.data = numberSpritesheet.data + digit * stride;
 		//this->drawSprite(sprite, xp, y, color);
