+++ Bullet.h
@@ -5,7 +5,7 @@
 class Bullet : public IDrawable{
 public:
 	int dir;
-	Bullet(size_t x, size_t y, int dir);
+	Bullet(int x, int y, int dir);
 	Sprite* getSprite();
 	void draw(Buffer * buffer);
 	void move();
