+++ IDrawable.h
@@ -7,7 +7,7 @@ class IDrawable
 {
 public:
 	uint32_t color = WHITE_UINT32;
-	size_t x, y, width, height;
+	int x, y, width, height;
 	virtual ~IDrawable() {};
 	virtual void draw(Buffer* buffer) = 0;
 };
